+++ CHANGELOG.md
@@ -5,10 +5,25 @@ All notable changes to this project from 1.1.0 forward will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-12-17
+
+### Removed
+
+-   Container Proxy are been removed
+    -   container.foo = 'foo'; // not works anymore
+    -   delete container.foo; // not works anymore
+    -   container.foo; // not works anymore
+
+### Fixed
+
+-   Internal assignement of parameters definition.type
+
 ## [1.1.1] - 2022-12-17
 
 ### Fixed
-- Internal assignement of parameters definition.type 
+
+-   Internal assignement of parameters definition.type
+
 ## [1.1.0] - 2022-12-17
 
 First Public Release On Npm